third_party/blink/web_tests/platform/generic/fast/forms/file/input-file-value-expected.txt

This tests the value attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS file.value is "C:\\fakepath\\foo.txt"
PASS file.files.length is 1
PASS file.value = 'foo' threw exception InvalidStateError: Failed to set the 'value' property on 'HTMLInputElement': This input element accepts a filename, which may only be programmatically set to the empty string..
PASS file.value is "C:\\fakepath\\foo.txt"
PASS file.files.length is 1
PASS file.value is ""
PASS file.files.length is 0
PASS successfullyParsed is true

TEST COMPLETE
